refactor: 리뷰사항인 CommonConstants 분리 및 패키지 변경
공통 부분 Common, 에러 부분 Error, 포맷 부분 Format으로 분리하였습니다.
