+++ src/main/java/com/codessquad/qna/constants/CommonConstants.java
@@ -1,0 +1,9 @@
+package com.codessquad.qna.constants;
+
+public class CommonConstants {
+
+    public static final String SESSION_LOGIN_USER = "loginUser";
+    public static final String REDIRECT_LOGIN_PAGE = "redirect:/users/login";
+
+    private CommonConstants() {}
+}
