+++ src/main/java/com/codessquad/qna/constants/ErrorConstants.java
@@ -1,0 +1,10 @@
+package com.codessquad.qna.constants;
+
+public class ErrorConstants {
+
+    public static final String ERROR_USER_NOT_FOUND = "error/user-not-found";
+    public static final String ERROR_CANNOT_EDIT_OTHER_USER_INFO = "error/cannot-edit-other-user-info";
+    public static final String ERROR_QUESTION_NOT_FOUND = "error/question-not-found";
+
+    private ErrorConstants() {}
+}
