+++ src/main/java/com/codessquad/qna/constants/FormatConstants.java
@@ -1,0 +1,11 @@
+package com.codessquad.qna.constants;
+
+import java.time.format.DateTimeFormatter;
+
+public class FormatConstants {
+
+    public static final DateTimeFormatter POST_DATA_DATE_TIME_FORMATTER =
+            DateTimeFormatter.ofPattern("YYYY-MM-dd HH:mm:ss");
+
+    private FormatConstants() {}
+}
