+++ src/main/java/com/codessquad/qna/error/ErrorController.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.error;
 
-import com.codessquad.qna.common.CommonConstants;
+import com.codessquad.qna.constants.ErrorConstants;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,16 +11,16 @@ public class ErrorController {
 
     @GetMapping("/user-not-found")
     public String goUserNotFoundPage() {
-        return CommonConstants.ERROR_USER_NOT_FOUND;
+        return ErrorConstants.ERROR_USER_NOT_FOUND;
     }
 
     @GetMapping("/question-not-found")
     public String goQuestionNotFoundPage() {
-        return CommonConstants.ERROR_QUESTION_NOT_FOUND;
+        return ErrorConstants.ERROR_QUESTION_NOT_FOUND;
     }
 
     @GetMapping("/cannot-edit-other-user-info")
     public String goCannotEditOtherUserInfoPage() {
-        return CommonConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
+        return ErrorConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
     }
 }
