+++ src/main/java/com/codessquad/qna/question/Question.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.question;
 
-import com.codessquad.qna.common.CommonConstants;
+import com.codessquad.qna.constants.FormatConstants;
 import com.codessquad.qna.user.User;
 
 import javax.persistence.*;
@@ -89,7 +89,7 @@ public class Question {
     }
 
     public String getFormattedCreatedDateTime() {
-        return createdDateTime.format(CommonConstants.POST_DATA_DATE_TIME_FORMATTER);
+        return createdDateTime.format(FormatConstants.POST_DATA_DATE_TIME_FORMATTER);
     }
 
     public LocalDateTime getUpdatedDateTime() {
@@ -101,7 +101,7 @@ public class Question {
     }
 
     public String getFormattedUpdatedDateTime() {
-        return updatedDateTime.format(CommonConstants.POST_DATA_DATE_TIME_FORMATTER);
+        return updatedDateTime.format(FormatConstants.POST_DATA_DATE_TIME_FORMATTER);
     }
 
     public boolean isDeleted() {
