+++ src/main/java/com/codessquad/qna/user/UserController.java
@@ -1,6 +1,7 @@
 package com.codessquad.qna.user;
 
-import com.codessquad.qna.common.CommonConstants;
+import com.codessquad.qna.constants.CommonConstants;
+import com.codessquad.qna.constants.ErrorConstants;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -47,7 +48,7 @@ public class UserController {
         try {
             modelAndView.addObject("user", getUserIfExist(id));
         } catch (NotFoundException e) {
-            return new ModelAndView(CommonConstants.ERROR_USER_NOT_FOUND);
+            return new ModelAndView(ErrorConstants.ERROR_USER_NOT_FOUND);
         }
         return modelAndView;
     }
@@ -58,11 +59,11 @@ public class UserController {
         try {
             User user = getUserIfExist(id);
             if (isLoginUserNotEqualsRequestedUser(session, user)) {
-                return CommonConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
+                return ErrorConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
             }
             model.addAttribute("user", user);
         } catch (NotFoundException e) {
-            return CommonConstants.ERROR_USER_NOT_FOUND;
+            return ErrorConstants.ERROR_USER_NOT_FOUND;
         }
         model.addAttribute("actionUrl", "/users/" + id);
         return "/users/modify-form";
@@ -76,7 +77,7 @@ public class UserController {
         try {
             User user = getUserIfExist(id);
             if (isLoginUserNotEqualsRequestedUser(session, user)) {
-                return CommonConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
+                return ErrorConstants.ERROR_CANNOT_EDIT_OTHER_USER_INFO;
             }
             if (user.isUserPasswordNotEquals(updateUser.getUserPassword())) {
                 return "redirect:/users/" + id + "/form";
@@ -84,7 +85,7 @@ public class UserController {
             user.update(updateUser, newPassword);
             session.setAttribute(CommonConstants.SESSION_LOGIN_USER, userRepository.save(user));
         } catch (NotFoundException e) {
-            return CommonConstants.ERROR_USER_NOT_FOUND;
+            return ErrorConstants.ERROR_USER_NOT_FOUND;
         } catch (TransactionSystemException e) {
             return "redirect:/users/" + id + "/form";
         }
